Move mysensors component to package
* Split code into multiple modules.
